+++ app/src/main/java/kt/school/starlord/App.kt
@@ -9,6 +9,7 @@ import org.koin.core.context.startKoin
 
 @SuppressLint("Registered")
 open class App : Application() {
+//    val pageLoader = PageLoader by inject()
 
     override fun onCreate() {
         super.onCreate()
@@ -20,7 +21,7 @@ open class App : Application() {
             androidLogger()
             androidContext(this@App)
             modules(
-                categoriesModule
+                    categoriesModule
             )
         }
     }
