+++ app/src/main/java/kt/school/starlord/di/CategoriesModule.kt
@@ -2,13 +2,16 @@ package kt.school.starlord.di
 
 import kt.school.starlord.domain.CategoriesRepository
 import kt.school.starlord.model.NetworkRepository
+import kt.school.starlord.model.PageLoader
 import kt.school.starlord.ui.categories.CategoriesViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val categoriesModule = module {
 
-    single<CategoriesRepository> { NetworkRepository() }
+    factory<PageLoader> {PageLoader()}
+
+    single<CategoriesRepository> { NetworkRepository(PageLoader()) }
 
     viewModel { CategoriesViewModel(get()) }
 }
