+++ app/src/main/java/kt/school/starlord/model/NetworkRepository.kt
@@ -1,13 +1,14 @@
 package kt.school.starlord.model
 
-import kt.school.starlord.di.PageLoader
-import kt.school.starlord.di.PageParser
+import android.app.Application
 import kt.school.starlord.domain.CategoriesRepository
 import kt.school.starlord.entity.Category
 
-class NetworkRepository : CategoriesRepository {
+
+/*inject module by constructor*/
+class NetworkRepository (private val pageLoader: PageLoader): CategoriesRepository {
     override fun getCategories(): List<Category> {
-        val page: String = PageLoader().page
+        val page: String = pageLoader.page
         val parser = PageParser()
         return parser.parsePage(page)
     }
