+++ app/src/main/java/kt/school/starlord/model/PageLoader.kt
@@ -1,0 +1,14 @@
+package kt.school.starlord.model
+
+import java.net.HttpURLConnection
+import java.net.URL
+
+class PageLoader() {
+        private val url: URL = URL("https://baraholka.onliner.by")
+        private val urlConnection: HttpURLConnection = url.openConnection() as HttpURLConnection
+        val page = try {
+            urlConnection.inputStream.bufferedReader().readText()
+        } finally {
+            urlConnection.disconnect()
+        }
+}
